Se realiza la modificación de varios formatos de campo en diversos módulos del sistema según lo indica el documento de la tarea que se adjunta a continuación

Related work items: #86475

+++ SITEL/GB.SIMEF.Entities/DetalleCategoriaNumerico.cs
@@ -20,8 +20,8 @@ namespace GB.SIMEF.Entities
         [Key]
         public int idCategoriaDetalle { get; set; }
         public int idCategoria { get; set; }
-        public int Minimo { get; set; }
-        public int Maximo { get; set; }
+        public double Minimo { get; set; }
+        public double Maximo { get; set; }
         public bool Estado { get; set; }
     }
 }
